+++ src/UI/DappModal/components/DappModal.tsx
@@ -2,6 +2,7 @@ import React, { ReactNode } from 'react';
 import classNames from 'classnames';
 import ReactDOM from 'react-dom';
 
+import { DataTestIdsEnum } from 'constants/index';
 import { WithClassnameType } from '../../types';
 import { DappModalConfig } from '../dappModal.types';
 import styles from '../dappModalStyles.scss';
@@ -32,7 +33,8 @@ export const DappModal = ({
   parentElement,
   config = defaultConfig,
   children,
-  className = 'dapp-modal-dialog-wrapper'
+  className = 'dapp-modal-dialog-wrapper',
+  'data-testid': dataTestId = DataTestIdsEnum.dappModal
 }: DappModalPropsType) => {
   if (!visible) {
     return null;
@@ -60,6 +62,7 @@ export const DappModal = ({
       role='dialog'
       aria-modal='true'
       className={classNames(modalDialogClassName, styles.dappModal, className)}
+      data-testid={dataTestId}
     >
       <div
         className={classNames(styles.dappModalContent, modalContentClassName)}