+++ src/UI/TransactionsToastList/components/TransactionToast/components/DefaultToastDeleteButton.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { WithClassnameType } from '../../../../types';
+import { DataTestIdsEnum } from 'constants/index';
+import { WithClassnameType } from 'types';
 
 export interface DefaultToastDeleteButtonPropsType extends WithClassnameType {
   onClick?: () => void;
@@ -17,7 +18,7 @@ export const DefaultToastDeleteButton = ({
       type='button'
       className={className}
       onClick={onClick}
-      data-testid='deleteToastButton'
+      data-testid={DataTestIdsEnum.deleteToastButton}
     >
       <FontAwesomeIcon icon={faTimes} size='xs' />
     </button>