+++ src/UI/walletConnect/WalletConnectLoginContainer/index.tsx
@@ -5,6 +5,7 @@ import QRCode from 'qrcode';
 
 import Lighting from 'assets/icons/lightning.svg';
 import globalStyles from 'assets/sass/main.scss';
+import { DataTestIdsEnum } from 'constants/index';
 import { useWalletConnectV2Login } from 'hooks/login/useWalletConnectV2Login';
 import { getAuthorizationInfo } from 'services/nativeAuth/helpers';
 import { ModalContainer } from 'UI/ModalContainer';
@@ -41,10 +42,10 @@ export const WalletConnectLoginContainer = ({
   callbackRoute,
   loginButtonText = 'xPortal App',
   title = 'Login with the xPortal App',
-  logoutRoute = '/unlock',
   className = 'dapp-wallet-connect-login-modal',
   lead = 'Scan the QR code using the xPortal App',
   wrapContentInsideModal = true,
+  logoutRoute,
   token,
   nativeAuth,
   onClose,
@@ -65,11 +66,11 @@ export const WalletConnectLoginContainer = ({
       wcPairings
     }
   ] = useWalletConnectV2Login({
-    logoutRoute,
     callbackRoute,
     token,
     nativeAuth,
-    onLoginRedirect
+    onLoginRedirect,
+    logoutRoute
   });
 
   const [qrCodeSvg, setQrCodeSvg] = useState<string>('');
@@ -206,7 +207,7 @@ export const WalletConnectLoginContainer = ({
         {isMobileDevice && (
           <a
             id='accessWalletBtn'
-            data-testid='accessWalletBtn'
+            data-testid={DataTestIdsEnum.accessWalletBtn}
             href={walletConnectDeepLinkV2}
             rel='noopener noreferrer nofollow'
             target='_blank'