+++ src/__mocks__/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './renderWithProvider';
 export * from './mockWindowLocation';
+export * from './mockWindowHistory';