+++ src/hooks/transactions/useSignTransactions.tsx
@@ -11,6 +11,7 @@ import {
   ERROR_SIGNING_TX,
   MISSING_PROVIDER_MESSAGE,
   PROVIDER_NOT_INITIALIZED,
+  SENDER_DIFFERENT_THAN_LOGGED_IN_ADDRESS,
   TRANSACTION_CANCELLED,
   WALLET_SIGN_SESSION
 } from 'constants/index';
@@ -56,7 +57,7 @@ export const useSignTransactions = () => {
   const providerType = getProviderType(provider);
   const isSigningRef = useRef(false);
   const setTransactionNonces = useSetTransactionNonces();
-  const { isGuarded } = useGetAccount();
+  const { isGuarded, address } = useGetAccount();
 
   const signTransactionsCancelMessage = useSelector(
     signTransactionsCancelMessageSelector
@@ -257,6 +258,18 @@ export const useSignTransactions = () => {
       return;
     }
 
+    const isLoggedInWithDifferentAccount = transactions.some((tx) => {
+      const sender = tx.getSender().toString();
+      return sender && address !== sender;
+    });
+
+    // Prevent signing transactions with different account
+    if (isLoggedInWithDifferentAccount) {
+      console.error(SENDER_DIFFERENT_THAN_LOGGED_IN_ADDRESS);
+
+      return onCancel(SENDER_DIFFERENT_THAN_LOGGED_IN_ADDRESS);
+    }
+
     /*
      * if the transaction is cancelled
      * the callback will go to undefined,
@@ -297,6 +310,7 @@ export const useSignTransactions = () => {
       console.error(errorMessage, err);
     }
   };
+
   useEffect(() => {
     if (hasTransactions) {
       signTransactions();